+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### 2.0.4 - UNRELEASED
 
 * Update Babylon preview to latest v3.1 preview
+* Add reflection maps to BabylonJS and ThreeJS previews
 
 ### 2.0.3 - 2017-06-12